pyproject.toml
[project]
name = "portals"
version = "0.1.0"
description = "Multi-platform document synchronization tool (Portals)"
authors = [
    {name = "Roman Siepelmeyer", email = "rsiepelmeyer@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "click>=8.1.7",
    "notion-client>=2.2.1",
    "watchdog>=4.0.0",
    "rich>=13.7.0",
    "structlog>=24.1.0",
    "pydantic>=2.5.0",
    "python-frontmatter>=1.1.0",
    "aiofiles>=23.2.1",
    "tenacity>=8.2.3",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.1.0",
    "markdown-it-py>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.2",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
    "types-aiofiles>=23.2.0",
]

[project.scripts]
docsync = "portals.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["portals"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

# Allow untyped decorators in CLI (Click decorators don't have full type stubs)
[[tool.mypy.overrides]]
module = "portals.cli.*"
disallow_untyped_decorators = false

# Allow untyped decorators in tests (pytest fixtures)
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_decorators = false

# Relax strict checks for utils (structlog has complex types)
[[tool.mypy.overrides]]
module = "portals.utils.*"
warn_return_any = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=portals --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["portals"]
omit = ["tests/*", "**/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
